Add email validation and user retrieval methods to UserService

+++ msvc-usuarios/src/main/java/com/osmar/springboot/di/app/msvcusuarios/services/UserServiceImpl.java
@@ -1,5 +1,8 @@
 package com.osmar.springboot.di.app.msvcusuarios.services;
 
+import com.osmar.springboot.di.app.msvcusuarios.clients.CursoClientRest;
+import com.osmar.springboot.di.app.msvcusuarios.exceptions.EmailAlreadyExistsException;
+import com.osmar.springboot.di.app.msvcusuarios.exceptions.UserNotFoundException;
 import com.osmar.springboot.di.app.msvcusuarios.models.entity.User;
 import com.osmar.springboot.di.app.msvcusuarios.repositories.UserRepository;
 import org.springframework.stereotype.Service;
@@ -15,14 +18,16 @@ public class UserServiceImpl implements UserService {
      * Repository for managing user entities.
      */
     private final UserRepository userRepository;
+    private final CursoClientRest cursoClientRest;
 
     /**
      * Constructor for UserServiceImpl.
      *
      * @param userRepository the repository to manage user entities
      */
-    public UserServiceImpl(UserRepository userRepository) {
+    public UserServiceImpl(UserRepository userRepository, CursoClientRest cursoClientRest) {
         this.userRepository = userRepository;
+        this.cursoClientRest = cursoClientRest;
     }
 
     @Override
@@ -40,7 +45,7 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional(readOnly = true)
     public Optional<User> findById(Long id) {
-        return userRepository.findById(id);
+        return Optional.ofNullable(userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id)));
     }
 
     /**
@@ -52,7 +57,36 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public User save(User user) {
+        userRepository.findByEmail(user.getEmail()).ifPresent(
+                existingUser -> {
+                    throw new EmailAlreadyExistsException(user.getEmail());
+                }
+        );
+
         return userRepository.save(user);
+    }
+
+
+    @Override
+    @Transactional
+    public User update(Long id, User user){
+        return userRepository.findById(id).map(
+                existingUser -> {
+                    userRepository.findByEmail(user.getEmail()).ifPresent(
+                            userWithEmail -> {
+                                if (!userWithEmail.getId().equals(id)) {
+                                    throw new EmailAlreadyExistsException(user.getEmail());
+                                }
+                            }
+                    );
+                    existingUser.setUsername(user.getUsername());
+                    existingUser.setEmail(user.getEmail());
+                    existingUser.setName(user.getName());
+                    existingUser.setLastName(user.getLastName());
+
+                    return userRepository.save(existingUser);
+                }
+        ).orElseThrow(() -> new UserNotFoundException(id));
     }
 
     /**
@@ -63,6 +97,34 @@ public class UserServiceImpl implements UserService {
     @Override
     @Transactional
     public void deleteById(Long id) {
+        if (!userRepository.existsById(id)) {
+            throw new UserNotFoundException(id);
+        }
         userRepository.deleteById(id);
+        cursoClientRest.deleteCursoUserById(id);
+    }
+
+    /**
+     * Finds a user by its email.
+     *
+     * @param email the email of the user to find
+     * @return an Optional containing the found user, or empty if not found
+     */
+    @Override
+    @Transactional(readOnly = true)
+    public Optional<User> findByEmail(String email) {
+        return userRepository.findByEmail(email);
+    }
+
+    /**
+     * Lists users by their IDs.
+     *
+     * @param ids the IDs of the users to list
+     * @return a list of users with the specified IDs
+     */
+    @Override
+    @Transactional(readOnly = true)
+    public List<User> listByIds(Iterable<Long> ids) {
+        return (List<User>) userRepository.findAllById(ids);
     }
 }
